Add translation layer for LibreHardwareMonitor enums
LibreHardwareMonitor enums make it into the final name for metrics. Thus
when an enum entry is renamed in an update, the metrics can be silently
updated, which can break carefully curated dashboards. The fix is to
introduce an intermediate translation layer to our own enums so that we
insulate any changes and can make a stronger guarantee to backwards
compatibility.