Add prometheus dashboard
The initial setup for OhmGraphite can be daunting if one has to create
their own dashboards from scratch. By creating a sample dashboard with
prometheus (others to follow), newcomers can be introduced to what types
of metrics are available. The dashboard is not perfect. It's not nearly
as compact as it could be, but it is general enough that others can use
it as a base.