Alias a sensor name by sensor id
LibreHardwareMonitor allows one to customize the name for a given
sensor. This customized name is stored in `LibreHardwareMonitor.config`
as a combination of sensor id and `/name` suffix:

```xml
<add key="/<sensor-id>/name" value="<alias>" />
```

Like the following:

```xml
<add key="/lpc/nct6792d/fan/1/name" value="superfan" />
```

The XML format of the config is the exact same as OhmGraphite's, so this commit
allows one to copy and paste an exact alias line from LibreHardwareMonitor's
config into OhmGraphite's config as shown below.

```xml
<?xml version="1.0" encoding="utf-8" ?>
<configuration>
  <appSettings>
    <add key="type" value="prometheus" />
    <add key="prometheus_port" value="4445" />
    <add key="prometheus_host" value="*" />

    <add key="/lpc/nct6792d/fan/1/name" value="superfan" />
   </appSettings>
</configuration>
```

The recommendation for users that want to alias is to prepare their
aliasing in LibreHardwareMonitor as that is the most surefire way to
generate the alias config (nothing beats copy + paste). For those not
interested in using LibreHardwareMonitor for creating aliases, InfluxDB
and Postgres / TimescaleDB users can craft it by hand by examining their
data sources as the sensor id is stored in those backends (Prometheus
collector omits sensor id due to avoid high cardinality and Graphite
translates it by removing inappropriate characters).

Aliasing sensors will allow users that additional ways to craft
dashboards with desired names as some backends make it difficult to
create aliases (eg: with postgres a `view` could be used but creating an
alias on a single metric in influxdb is more difficult).