Include sensor name when sensor is added
With sensor name aliasing on master there should be a self contained way
to determine the sensor id for a given name, as relying on users running
LibreHardwareMonitor would be too inconvenient.

This commit modifies the "Sensor added" log line, which already prints
out the sensor's id to include the sensor name.

Now if someone wants to update the "Fan #2" sensor they can open the
OhmGraphite.log and search for "Fan #2" and see the following line:

```
Sensor added: /lpc/nct6792d/fan/1 "Fan #2"
```

Then the this line can be used to add the alias line to
`OhmGraphite.exe.config`

```xml
<add key="/lpc/nct6792d/fan/1/name" value="superfan" />
```