Upgrade to .NET 5.0
By moving to .NET 5.0 we can leverage the builtin build system to make a
single executable instead of Fody and ilk. We now create a self
contained executable instead of relying on the user having the .net 5.0
runtime installed (which I'm assuming will be a rarity for quite some
time). The biggest downside is the executable size is larger 10x larger
(3MB vs 30MB).