Configurable certificate verification
This commit introduces the `certificate_verification` property:

```xml
<?xml version="1.0" encoding="utf-8" ?>
<configuration>
  <appSettings>
    <add key="certificate_verification"
         value="C:\apps\OhmGraphite\influxdb-selfsigned.crt" />
   </appSettings>
</configuration>
```

The possible values:

 - True (the default): all certificates are verified
 - False: No certificates are verified (INSECURE)
 - a file path of a certificate that the server is allowed to return and
 still be considered a valid request (useful for self signed
 certificates). Recommended to be an absolute file path.

The nice thing about targeting .net 5 is that reading of pem
certificates is baked in (as well as other formats), so one can take the
output from an openssl command and copy the file here.

The downside of this commit is that it does now include more code so
that the executable is larger (33MB to 39MB) and it takes about that
much RAM as well.