Capture additional components in prometheus sample dashboard
When querying the breakdown of hardware components, the prometheus query may
be too narrow and exclude components that didn't match but should have been
included. For instance the previous hardware query:

```
ohm_hdd_load_percent{instance="$instance"}
```

Will exclude hard drives without load percentages.

The fix is to match on name prefix, so anything that starts with `ohm_hdd_` is
considered a harddrive and will have its own details section.

Grafana version was updated in the meantime, as there are still some
UI oddities in the dashboard remaining, which may be fixed in a future
update