Use 100 °C as soft max for temperature panels in sample dashboard
It's preferable to have a at a glance view of temperatures panels to
know which one is hot. By setting the soft max value they will all share
the same y-axis. One small difference is that soft max is preferred to
max as temperatures may exceed 100 degrees.

As for why 100 degrees was chosen: somewhat arbitrary. It's a nice round
number that components throttle around