Remove threshold lines from cpu temp sample panel
None of the other temperature panels contained thresholds and it didn't
seem right to add threshold to others as the threshold palette clashed
with the series palette. And I don't have the intuition to make this
color choice so the easiest thing to do is to remove it. I don't think
all is lost as since the soft max is 100 degrees, if any graph shows a
high reading it would be interpreted the same as crossing a threshold.