Add version option when interactively running
This adds the ability to query OhmGraphite for it's version.

```
.\OhmGraphite.exe run --version
```

The above will print something like the following, with the version printed on
the last line:

```
2021-11-06 07:34:26.0893|INFO|Topshelf.HostFactory|Configuration Result:
[Success] Name OhmGraphite
[Success] DisplayName Ohm Graphite
[Success] Description Extract hardware sensor data and exports it to a given host and port in a graphite compatible format
[Success] ServiceName OhmGraphite
2021-11-06 07:34:26.1141|INFO|Topshelf.HostConfigurators.HostConfiguratorImpl|Topshelf v4.3.0.0, .NET 5.0.6 (5.0.6)
0.22.0.0
```

This serves an alternative to the previous ways of determining the version
installed with powershell:

```pwsh
(Get-Item .\OhmGraphite.exe).VersionInfo.FileVersion
```

or

```pwsh
(Get-Command .\OhmGraphite.exe).Version
```

or right clicking on the executable to get details.